pyproject.toml
[tool.poetry]
name = "ontokit"
version = "0.1.0"
description = ""
authors = ["Felix-Brandstetter <Felix.Brandstetter@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.36.1"
python-dotenv = "^1.0.0"
linkml = "^1.6.7"
jinja2 = "^3.1.2"
torch = "^2.1.2"
guidance = "^0.1.10"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
pylint = "^3.0.3"
pytest = "^7.4.4"
accelerate = "^0.24.0"
torch = {url = "https://download.pytorch.org/whl/cu121/torch-2.1.2%2Bcu121-cp311-cp311-win_amd64.whl"}
bitsandbytes = {url = "https://github.com/jllllll/bitsandbytes-windows-webui/releases/download/wheels/bitsandbytes-0.41.0-py3-none-win_amd64.whl"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
